+++ src/gui/explorer.go
@@ -229,7 +229,7 @@ func getRichlist(gateway Gatewayer) http.HandlerFunc {
 		var topn int
 		topnStr := r.FormValue("n")
 		if topnStr == "" {
-			topn = 0
+			topn = 20
 		} else {
 			var err error
 			topn, err = strconv.Atoi(topnStr)