+++ src/gui/integration/integration_test.go
@@ -1,6 +1,7 @@
 package gui_integration_test
 
 import (
+	"encoding/hex"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -11,9 +12,11 @@ import (
 	"sort"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/daemon"
 	"github.com/skycoin/skycoin/src/gui"
 	"github.com/skycoin/skycoin/src/util/droplet"
@@ -138,9 +141,10 @@ func loadGoldenFile(t *testing.T, filename string, testData TestData) {
 }
 
 func updateGoldenFile(t *testing.T, filename string, content interface{}) {
-	contentJson, err := json.MarshalIndent(content, "", "\t")
+	contentJSON, err := json.MarshalIndent(content, "", "\t")
 	require.NoError(t, err)
-	err = ioutil.WriteFile(filename, contentJson, 0644)
+	contentJSON = append(contentJSON, '\n')
+	err = ioutil.WriteFile(filename, contentJSON, 0644)
 	require.NoError(t, err)
 }
 
@@ -1067,7 +1071,7 @@ func TestLiveTransaction(t *testing.T) {
 			}
 			var expected *visor.ReadableTransaction
 			loadGoldenFile(t, tc.goldenFile, TestData{tx, &expected})
-			require.Equal(t, expected, tx)
+			require.Equal(t, expected, &tx.Transaction)
 		})
 	}
 }
@@ -1131,7 +1135,7 @@ func TestStableTransaction(t *testing.T) {
 
 			var expected *visor.ReadableTransaction
 			loadGoldenFile(t, tc.goldenFile, TestData{tx, &expected})
-			require.Equal(t, expected, tx)
+			require.Equal(t, expected, &tx.Transaction)
 		})
 	}
 }
@@ -1595,7 +1599,22 @@ func TestStableAddressTransactions(t *testing.T) {
 		},
 	}
 
-	testAddressTransactions(t, cases)
+	c := gui.NewClient(nodeAddress())
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			txns, err := c.AddressTransactions(tc.address)
+			if tc.errMsg != "" {
+				assertResponseError(t, err, tc.errCode, tc.errMsg)
+				return
+			}
+
+			require.NoError(t, err)
+
+			var expected []gui.ReadableTransaction
+			loadGoldenFile(t, tc.golden, TestData{txns, &expected})
+			require.Equal(t, expected, txns)
+		})
+	}
 }
 
 func TestLiveAddressTransactions(t *testing.T) {
@@ -1627,11 +1646,10 @@ func TestLiveAddressTransactions(t *testing.T) {
 		},
 	}
 
-	testAddressTransactions(t, cases)
-}
-
-func testAddressTransactions(t *testing.T, cases []addressTransactionsTestCase) {
 	c := gui.NewClient(nodeAddress())
+	// Get current blockchain height
+	bp, err := c.BlockchainProgress()
+	require.NoError(t, err)
 	for _, tc := range cases {
 		t.Run(tc.name, func(t *testing.T) {
 			txns, err := c.AddressTransactions(tc.address)
@@ -1644,6 +1662,12 @@ func testAddressTransactions(t *testing.T, cases []addressTransactionsTestCase) 
 
 			var expected []gui.ReadableTransaction
 			loadGoldenFile(t, tc.golden, TestData{txns, &expected})
+
+			// Recaculate the height if it's live test
+			for i := range expected {
+				expected[i].Status.Height = bp.Current - expected[i].Status.BlockSeq + 1
+			}
+
 			require.Equal(t, expected, txns)
 		})
 	}
@@ -1774,4 +1798,482 @@ func TestLivePendingTransactions(t *testing.T) {
 
 	_, err := c.PendingTransactions()
 	require.NoError(t, err)
+}
+
+func TestLiveWalletSpend(t *testing.T) {
+	if !doLive(t) {
+		return
+	}
+
+	doLiveEnvCheck(t)
+
+	c := gui.NewClient(nodeAddress())
+	w, totalCoins, _ := prepareAndCheckWallet(t, c, 2e6, 2)
+	tt := []struct {
+		name    string
+		to      string
+		coins   uint64
+		errMsg  []byte
+		checkTx func(t *testing.T, tx *visor.TransactionResult)
+	}{
+		{
+			name:  "send all coins to the first address",
+			to:    w.Entries[0].Address.String(),
+			coins: totalCoins,
+			checkTx: func(t *testing.T, tx *visor.TransactionResult) {
+				// Confirms the total output coins are equal to the totalCoins
+				var coins uint64
+				for _, o := range tx.Transaction.Out {
+					c, err := droplet.FromString(o.Coins)
+					require.NoError(t, err)
+					coins += c
+				}
+				require.Equal(t, totalCoins, coins)
+
+				// Confirms the address balance are equal to the totoalCoins
+				coins, _ = getAddressBalance(t, c, w.Entries[0].Address.String())
+				require.Equal(t, totalCoins, coins)
+			},
+		},
+		{
+			// send 0.001 coin to the second address,
+			name:  "send 0.001 coin to second address",
+			to:    w.Entries[1].Address.String(),
+			coins: 1e3,
+			checkTx: func(t *testing.T, tx *visor.TransactionResult) {
+				// Confirms there're two outputs, one to the second address, one as change output to the first address.
+				require.Len(t, tx.Transaction.Out, 2)
+
+				// Gets the output of the second address in the transaction
+				getAddrOutputInTx := func(t *testing.T, tx *visor.TransactionResult, addr string) *visor.ReadableTransactionOutput {
+					for _, output := range tx.Transaction.Out {
+						if output.Address == addr {
+							return &output
+						}
+					}
+					t.Fatalf("transaction doesn't have output to address: %v", addr)
+					return nil
+				}
+
+				out := getAddrOutputInTx(t, tx, w.Entries[1].Address.String())
+
+				// Confirms the second address has 0.001 coin
+				require.Equal(t, out.Coins, "0.001000")
+				require.Equal(t, out.Address, w.Entries[1].Address.String())
+
+				coin, err := droplet.FromString(out.Coins)
+				require.NoError(t, err)
+
+				// Gets the expected change coins
+				expectChangeCoins := totalCoins - coin
+
+				// Gets the real change coins
+				changeOut := getAddrOutputInTx(t, tx, w.Entries[0].Address.String())
+				changeCoins, err := droplet.FromString(changeOut.Coins)
+				require.NoError(t, err)
+				// Confirms the change coins are matched.
+				require.Equal(t, expectChangeCoins, changeCoins)
+			},
+		},
+	}
+
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			result, err := c.Spend(w.GetFilename(), tc.to, tc.coins)
+			if err != nil {
+				t.Fatalf("spend failed: %v", err)
+			}
+
+			tk := time.NewTicker(time.Second)
+			var tx *visor.TransactionResult
+		loop:
+			for {
+				select {
+				case <-time.After(30 * time.Second):
+					t.Fatal("Waiting for transaction to be confirmed timeout")
+				case <-tk.C:
+					tx = getTransaction(t, c, result.Transaction.Hash)
+					if tx.Status.Confirmed {
+						break loop
+					}
+				}
+			}
+			tc.checkTx(t, tx)
+		})
+	}
+
+	// Confirms sending coins less than 0.001 is not allowed
+	errMsg := "500 Internal Server Error - Transaction violates soft constraint: invalid amount, too many decimal places\n"
+	for i := uint64(1); i < uint64(1000); i++ {
+		cs, err := droplet.ToString(i)
+		require.NoError(t, err)
+		name := fmt.Sprintf("send invalid coin %v", cs)
+		t.Run(name, func(t *testing.T) {
+			result, err := c.Spend(w.GetFilename(), w.Entries[0].Address.String(), i)
+			require.Equal(t, errMsg, err.Error())
+			require.Nil(t, result)
+		})
+	}
+}
+
+func TestCreateWallet(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+
+	w, clean := createWallet(t, c)
+	defer clean()
+
+	walletDir := getWalletDirFromEnv(t)
+
+	// Confirms the wallet does exist
+	walletPath := filepath.Join(walletDir, w.GetFilename())
+	_, err := os.Stat(walletPath)
+	require.NoError(t, err)
+
+	checkWalletEntriesAndLastSeed(t, w)
+}
+
+func TestGetWallet(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+
+	// Create a wallet
+	w, clean := createWallet(t, c)
+	defer clean()
+
+	// Confirms the wallet can be acquired
+	w1, err := c.Wallet(w.GetFilename())
+	require.NoError(t, err)
+	require.Equal(t, *w, *w1)
+}
+
+func TestGetWallets(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+
+	// Creates 2 new wallets
+	var ws []wallet.Wallet
+	for i := 0; i < 2; i++ {
+		w, clean := createWallet(t, c)
+		defer clean()
+		// cleaners = append(cleaners, clean)
+		ws = append(ws, *w)
+	}
+
+	// Gets wallet from node
+	readableWallets, err := c.Wallets()
+	require.NoError(t, err)
+
+	// Create the wallet map
+	walletMap := make(map[string]wallet.Wallet)
+	for _, rw := range readableWallets {
+		w, err := rw.ToWallet()
+		require.NoError(t, err)
+		walletMap[w.GetFilename()] = w
+	}
+
+	// Confirms the returned wallets contains the wallet we created.
+	for _, w := range ws {
+		retW, ok := walletMap[w.GetFilename()]
+		require.True(t, ok)
+		require.Equal(t, w, retW)
+	}
+}
+
+// TestWalletNewAddress will generate 30 wallets for testing, and they will
+// be removed automatically after testing.
+//
+// Note: Though the new generated wallet files are all deleted, they are still
+// in the memory, you need to resrat the node to free them.
+func TestWalletNewAddress(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+	// We only test 30 cases, cause the more addresses we generate, the longer
+	// it takes, we don't want to spend much time here.
+	for i := 1; i <= 30; i++ {
+		name := fmt.Sprintf("generate %v addresses", i)
+		t.Run(name, func(t *testing.T) {
+			w, clean := createWallet(t, c)
+			defer clean()
+
+			addrs, err := c.NewWalletAddress(w.GetFilename(), i)
+			if err != nil {
+				t.Fatalf("%v", err)
+				return
+			}
+			require.NoError(t, err)
+
+			as, err := w.GenerateAddresses(uint64(i))
+			require.NoError(t, err)
+
+			// Confirms thoses new generated addresses are the same.
+			require.Equal(t, len(addrs), len(as))
+			for i, a := range as {
+				require.Equal(t, a.String(), addrs[i])
+			}
+		})
+	}
+}
+
+func TestStableWalletBalance(t *testing.T) {
+	if !doStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+	w, clean := createWallet(t, c)
+	defer clean()
+
+	bp, err := c.WalletBalance(w.GetFilename())
+	require.NoError(t, err)
+
+	var expect wallet.BalancePair
+	loadGoldenFile(t, "wallet-balance.golden", TestData{bp, &expect})
+	require.Equal(t, expect, *bp)
+}
+
+func TestLiveWalletbalance(t *testing.T) {
+	if !doLive(t) {
+		return
+	}
+
+	doLiveEnvCheck(t)
+
+	_, walletName := getWalletFromEnv(t)
+	c := gui.NewClient(nodeAddress())
+	bp, err := c.WalletBalance(walletName)
+	require.NoError(t, err)
+	require.NotNil(t, bp)
+}
+
+func TestWalletUpdate(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+	w, clean := createWallet(t, c)
+	defer clean()
+
+	err := c.UpdateWallet(w.GetFilename(), "new wallet")
+	require.NoError(t, err)
+
+	// Confirms the wallet has label of "new wallet"
+	w1, err := c.Wallet(w.GetFilename())
+	require.NoError(t, err)
+	require.Equal(t, w1.GetLabel(), "new wallet")
+}
+
+func TestStableWalletTransactions(t *testing.T) {
+	if !doStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+	w, clean := createWallet(t, c)
+	defer clean()
+
+	txns, err := c.WalletTransactions(w.GetFilename())
+	require.NoError(t, err)
+
+	var expect gui.UnconfirmedTxnsResponse
+	loadGoldenFile(t, "wallet-transactions.golden", TestData{txns, &expect})
+	require.Equal(t, expect, *txns)
+}
+
+func TestLiveWalletTransactions(t *testing.T) {
+	if !doLive(t) {
+		return
+	}
+
+	doLiveEnvCheck(t)
+
+	c := gui.NewClient(nodeAddress())
+	w, _, _ := prepareAndCheckWallet(t, c, 1e6, 1)
+	txns, err := c.WalletTransactions(w.GetFilename())
+	require.NoError(t, err)
+
+	bp, err := c.WalletBalance(w.GetFilename())
+	require.NoError(t, err)
+	// There's pending transactions if predicted coins are not the same as confirmed coins
+	if bp.Predicted.Coins != bp.Confirmed.Coins {
+		require.NotEmpty(t, txns.Transactions)
+		return
+	}
+
+	require.Empty(t, txns.Transactions)
+}
+
+func TestWalletFolderName(t *testing.T) {
+	if !doLiveOrStable(t) {
+		return
+	}
+
+	c := gui.NewClient(nodeAddress())
+	folderName, err := c.WalletFolderName()
+	require.NoError(t, err)
+
+	walletDir := getWalletDirFromEnv(t)
+	require.Equal(t, walletDir, folderName.Address)
+}
+
+// prepareAndCheckWallet gets wallet from environment, and confirms:
+// 1. The minimal coins and coin hours requirements are met.
+// 2. The wallet has at least two address entry.
+// Returns the loaded wallet, total coins and total coin hours in the wallet.
+func prepareAndCheckWallet(t *testing.T, c *gui.Client, miniCoins, miniCoinHours uint64) (*wallet.Wallet, uint64, uint64) {
+	walletDir, walletName := getWalletFromEnv(t)
+	walletPath := filepath.Join(walletDir, walletName)
+
+	// Checks if the wallet does exist
+	if _, err := os.Stat(walletPath); os.IsNotExist(err) {
+		t.Fatalf("Wallet %v doesn't exist", walletPath)
+	}
+
+	w, err := wallet.Load(walletPath)
+	if err != nil {
+		t.Fatalf("Load wallet %v failed: %v", walletPath, err)
+	}
+
+	// Generate more addresses if address entries less than 2.
+	if len(w.Entries) < 2 {
+		_, err := c.NewWalletAddress(w.GetFilename(), 2-len(w.Entries))
+		if err != nil {
+			t.Fatalf("New wallet address failed: %v", err)
+		}
+	}
+
+	coins, hours := getWalletBalance(t, c, walletName)
+	if coins < miniCoins {
+		t.Fatalf("Wallet must have at least %d coins", miniCoins)
+	}
+
+	if hours < miniCoinHours {
+		t.Fatalf("Wallet must have at least %d coin hours", miniCoinHours)
+	}
+
+	if err := w.Save(walletDir); err != nil {
+		t.Fatalf("%v", err)
+	}
+
+	return w, coins, hours
+}
+
+// getWalletFromEnv loads wallet from envrionment variables.
+// Returns wallet dir and wallet name.
+func getWalletFromEnv(t *testing.T) (string, string) {
+	walletDir := os.Getenv("WALLET_DIR")
+	if walletDir == "" {
+		t.Fatal("Missing WALLET_DIR environment value")
+	}
+
+	walletName := os.Getenv("WALLET_NAME")
+	if walletName == "" {
+		t.Fatal("Missing WALLET_NAME environment value")
+	}
+
+	return walletDir, walletName
+}
+
+func getWalletDirFromEnv(t *testing.T) string {
+	walletDir := os.Getenv("WALLET_DIR")
+	if walletDir == "" {
+		t.Fatal("Missing WALLET_DIR environment value")
+	}
+	return walletDir
+}
+
+func doLiveEnvCheck(t *testing.T) {
+	t.Helper()
+	walletDir := os.Getenv("WALLET_DIR")
+	if walletDir == "" {
+		t.Fatal("missing WALLET_DIR environment value")
+	}
+
+	walletName := os.Getenv("WALLET_NAME")
+	if walletName == "" {
+		t.Fatal("missing WALLET_NAME environment value")
+	}
+}
+
+// getWalletBalance gets wallet balance.
+// Returns coins and hours
+func getWalletBalance(t *testing.T, c *gui.Client, walletName string) (uint64, uint64) {
+	wp, err := c.WalletBalance(walletName)
+	if err != nil {
+		t.Fatalf("Get wallet balance of %v failed: %v", walletName, err)
+	}
+
+	return wp.Confirmed.Coins, wp.Confirmed.Hours
+}
+
+func getTransaction(t *testing.T, c *gui.Client, txid string) *visor.TransactionResult {
+	tx, err := c.Transaction(txid)
+	if err != nil {
+		t.Fatalf("%v", err)
+	}
+
+	return tx
+}
+
+// getAddressBalance gets balance of given address.
+// Returns coins and coin hours.
+func getAddressBalance(t *testing.T, c *gui.Client, addr string) (uint64, uint64) {
+	bp, err := c.Balance([]string{addr})
+	if err != nil {
+		t.Fatalf("%v", err)
+	}
+	return bp.Confirmed.Coins, bp.Confirmed.Hours
+}
+
+// checkWalletEntriesAndLastSeed confirms the wallet entries and lastSeed are derivied
+// from the seed.
+func checkWalletEntriesAndLastSeed(t *testing.T, w *wallet.Wallet) {
+	seed, ok := w.Meta["seed"]
+	require.True(t, ok)
+	newSeed, seckeys := cipher.GenerateDeterministicKeyPairsSeed([]byte(seed), len(w.Entries))
+	require.Len(t, seckeys, len(w.Entries))
+	for i, sk := range seckeys {
+		require.Equal(t, w.Entries[i].Secret, sk)
+		pk := cipher.PubKeyFromSecKey(sk)
+		require.Equal(t, w.Entries[i].Public, pk)
+	}
+	lastSeed, ok := w.Meta["lastSeed"]
+	require.True(t, ok)
+	require.Equal(t, lastSeed, hex.EncodeToString(newSeed))
+}
+
+// createWallet creates a wallet with rand seed.
+// Returns the generated wallet and clean up function.
+func createWallet(t *testing.T, c *gui.Client) (*wallet.Wallet, func()) {
+	seed := hex.EncodeToString(cipher.RandByte(32))
+	// Use the first 6 letter of the seed as label.
+	rw, err := c.CreateWallet(seed, seed[:6], 0)
+	require.NoError(t, err)
+
+	w, err := rw.ToWallet()
+	require.NoError(t, err)
+	err = w.Validate()
+	require.NoError(t, err)
+
+	walletDir := getWalletDirFromEnv(t)
+
+	return &w, func() {
+		// Cleaner function to delete the wallet
+		walletPath := filepath.Join(walletDir, w.GetFilename())
+		err = os.Remove(walletPath)
+		require.NoError(t, err)
+	}
 }