+++ src/gui/wallet_test.go
@@ -789,6 +789,7 @@ func TestWalletTransactionsHandler(t *testing.T) {
 		WalletID string
 	}
 
+	unconfirmedTxn, _ := visor.NewReadableUnconfirmedTxn(&visor.UnconfirmedTxn{})
 	tt := []struct {
 		name                                  string
 		method                                string
@@ -798,7 +799,7 @@ func TestWalletTransactionsHandler(t *testing.T) {
 		walletID                              string
 		gatewayGetWalletUnconfirmedTxnsResult []visor.UnconfirmedTxn
 		gatewayGetWalletUnconfirmedTxnsErr    error
-		responseBody                          []visor.UnconfirmedTxn
+		responseBody                          UnconfirmedTxnsResponse
 	}{
 		{
 			name:   "405",
@@ -855,7 +856,7 @@ func TestWalletTransactionsHandler(t *testing.T) {
 			err:      "",
 			walletID: "foo",
 			gatewayGetWalletUnconfirmedTxnsResult: make([]visor.UnconfirmedTxn, 1),
-			responseBody:                          []visor.UnconfirmedTxn{visor.UnconfirmedTxn{}},
+			responseBody:                          UnconfirmedTxnsResponse{Transactions: []visor.ReadableUnconfirmedTxn{*unconfirmedTxn}},
 		},
 	}
 
@@ -895,10 +896,13 @@ func TestWalletTransactionsHandler(t *testing.T) {
 			require.Equal(t, tc.err, strings.TrimSpace(rr.Body.String()), "case: %s, handler returned wrong error message: got `%v`| %s, want `%v`",
 				tc.name, strings.TrimSpace(rr.Body.String()), status, tc.err)
 		} else {
-			var msg []visor.UnconfirmedTxn
+			var msg UnconfirmedTxnsResponse
 			err = json.Unmarshal(rr.Body.Bytes(), &msg)
 			require.NoError(t, err)
-			require.Equal(t, tc.responseBody, msg, tc.name)
+			// require.Equal on whole response might result in flaky tests as there is a time field attached to unconfirmed txn response
+			require.IsType(t, msg, tc.responseBody)
+			require.Len(t, msg.Transactions, 1)
+			require.Equal(t, msg.Transactions[0].Txn, tc.responseBody.Transactions[0].Txn)
 		}
 	}
 }
@@ -909,7 +913,6 @@ func TestWalletCreateHandler(t *testing.T) {
 		Label string
 		ScanN string
 	}
-
 	tt := []struct {
 		name                      string
 		method                    string
@@ -1147,7 +1150,6 @@ func TestWalletNewSeed(t *testing.T) {
 	type httpBody struct {
 		Entropy string
 	}
-
 	tt := []struct {
 		name      string
 		method    string
@@ -1274,6 +1276,7 @@ func TestWalletNewAddressesHandler(t *testing.T) {
 	type Addresses struct {
 		Address []string `json:"addresses"`
 	}
+
 	var responseAddresses = Addresses{}
 	var responseEmptyAddresses = Addresses{}